SendSS: ! fixed non-centered text in capture desktop size edit box (resource.rc) ! fixed x86_64 related type error (CSendImageShack) - removed duplicate code, and "removed" ability to use custom icons for capture tabs (we're using the same icon anyway, no need to load it thrice) + added local file feature to upload existing local files (resource, UMainForm, mir_string)   reduced function DlgProc_CaptureWindow and DlgProc_CaptureDesktop to just DlgProc_CaptureTabPage as we use less code with only one function handling all 3 tab pages ("lot" of shared code)
git-svn-id: http://svn.miranda-ng.org/main/trunk@9416 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SendScreenshotPlus/src/CSendImageShack.cpp
@@ -79,7 +79,7 @@ void CSendImageShack::SendThread() {
 			if(url && *url){
 				mir_free(m_URL), m_URL=mir_strdup(url);
 				mir_free(m_URLthumb), m_URLthumb=mir_strdup(m_URL);
-				int extlen;
+				size_t extlen;
 				char* pos=strrchr(m_URLthumb,'.');
 				if(pos && (extlen=mir_strlen(pos))>2){
 					char* tmp=mir_strdup(pos);
